transactions: Split cancellation and state retrieval
Currently, all state about a transaction is reported on transaction
cancellation. While this is the natural point in time when it should
happen, it is an awkward interface to have the state be returned by the
cancellation function. So let's split up cancellation of transactions
and state retrieval such that callers interestend in the state need to
call `transaction.State()` after the transaction was cancelled.