transactions: Do not return state when cancelling
The transaction state is currently returned whenever one calls the
transaction cleanup function. This is an artifact caused by the
transaction type not being publicly visible earlier, but it starts to
become a hinderance when implementing new features and querying
transacton state. So let's just stop returning the state from the
cleanup function and ask callers to retrieve the state by themselves.