datastore: Fix unsynced Goroutine causing racy test case
One of the datastore tests verifies that if we trigger the health
updater twice, that the job lock's timestamp got updated. The way we
run the health updater is racy, though, as we do not verify that it
actually exits before the test case finishes. As a result, it may
execute queries against the database while the follwing test truncates
all tables already.

Let's fix this race by synchronizing shutdown of the Goroutine.