praefect: Fix service registration after start
It's not allowed to register any more services with a GRPC server after
it's been started already. But due to a race, we may do exactly that
when registering new mocked SmartHTTP servers: `NewServerWithHealth()`
not only creates a new GRPC server with the health service registered,
but already starts it. As we still need to register the SmartHTTP server
afterwards, we're thus racing against the server's `Serve()` Goroutine.

Fix this race by assembling the server manually before calling
`Serve()` on it.