datastore: Fix detection of context cancellation in health update
When the context for `StartHealthUpdate()` got canceled, we do not want
to return an error but instead just stop the health updating loop and
return. We need to treat context cancellation specially when executing
SQL queries, though, as the pq driver will not return `context.Canceled`
or `context.DeadlineExceeded` errors, but instead a custom error code
"query_canceled".

Fix detection of context cancellation by checking for any `pq.Error`s
with the mentioned error code.