git: Add function to check for reference-transaction hooks
The new reference-transaction hook has been introduced with Git v2.28.0.
As that release is currently in a pre-release state and as our minimum
required Git version is thus lower than v2.28.0, we cannot assume the
hook to be supported. This commit thus adds a new function to check
whether the hook is supported or not.