Makefile: build golangci-lint via Go directly
To enable use of golangci-lint, we download pre-compiled binaries from
the respective release. While this is comparatively quick, it requires
us to use a per-platform switch that downloads the correct set of
binaries for the host platform. This is a bit unwieldy and only allows
us to support platforms for which golang-ci lint itself provides
pre-built binaries.

Improve the situation by just building golangci-lint via `go get` like
we do with other tools.