Makefile: Do not error out early on unsupported platforms
We currently treat only macOS and Linux as supported platforms for
building Gitaly and error out early on all other platforms. This is
needlessly strict, though, as most of the build process actually works
just fine on other platforms, too.

One exception of this is generating sources via protoc. As we're
downloading protoc as a pre-built binary, we can only support platforms
for which upstream provides any binaries. But given that protoc isn't
required by integrators, or even by developers in most cases, it's
needlessly strict to abort any build just because we can't get hold of a
protoc binary that we don't even require in the first place. So let's
stop erroring out early and only abort the build when trying to set up
protoc.