cmd: Add gitaly-git2go command
This adds a new gitaly-git2go command that is currently intended as a
proof-of-concept for our Git2Go build framework. It thus doesn't do much
except demonstrating it's possible to link against Git2Go and its
bundled lbigit2 static archive and use its functionality. At a later
point, this is going to be the entrypoint for any Git functionality that
we want to implement via Git2Go.