coordinator: Extract function to register transactions
Transaction registration is handled in-line when assembling mutator
stream parameters. Given that this function is long enough already and
that we're about to allow for different voting strategies which will
require us to extend the transaction registration logic, let's pull it
out into a separate function to make it easier to extend.