add accept-dataloss sub-command
Adds 'accept-dataloss' sub-command which allows designating a new
authoritative storage for a repository. This might be necessary if
a repository is in read-only mode and recovering the data is not
viable. The current version of the repository on the authoritative
storage is considered latest and versions on the other physical
storages will be overwritten with it.