transactions: Always include all voters in results
The transaction outcome of whether a voter has succeeded in all votes or
not is currently created by iterating over all subtransactions. As a
result, the outcome will be completely empty if no subtransaction exists
at all, which makes it less useful in that case.

Let's improve this situation by seeding the result map with all voters
part of the transaction in case there are no subtransactions.