metric for the number of read-only repositories
Adds a Prometheus collector for RepositoryStore to expose the number
of read-only repositories within a virtual storage.

Since the state of the repositories is in the database, the current
approach leads to redundant work as each Praefect will query the
database for the number of read-only repositories. This could be
fixed later by extracting an exporter from the metrics that use the
database as their source of truth.

Alternatively, synchronization between Praefect nodes could be added
in order to coordinate which Praefect should perform the potentially
expensive queries for metrics.