smarthttp: Fix PostReceive reftx hook test
While TestPostReceiveWithReferenceTransactionHook succeeds when run as
part of the smarthttp testsuite, it fails when run standalone. The issue
is that the test didn't set up the hooks override and thus hooks cannot
be located correctly in case the test is run standalone. In case it's
run as part of the smarthttp suite, it succeeds because previous tests
failed to correctly restore the old value.

Fix the test by correctly setting up `hooks.Override` and correctly
restoring its value after the tests.