ci: Cache libgit2 library
The new libgit2 build target will be run by CI jobs in multiple stages.
As builds require a non-negligible amount of time, it adds up and thus
slows down our CI pipeline.

Let's work against this by using a file-keyed cache. As it's impossible
to use multiple caches, we also have to convert the existing cache for
Ruby to be. The result is thus keyed by both the Gemfile.lock and
Makefile, as both are definitive sources for what's being built.