Makefile: Rebuild versioned dependencies on changes
Currently, versioned dependencies like protoc, golangci-lint et al. will
get built once only. This is desirable most of the time, except for the
case where the dependency's version itself or its build instructions
changed. In such a case, we'll end up with a stale version of the
dependency as it doesn't get updated at all.

Fix this issue by adding a dependency on the Makefile itself. If it
changes, then it may indicate that the versioned dependency has changed
too and will thus require a rebuild. While this is a tad rough and will
cause us to build those dependencies even in cases where they don't
require a rebuilt, it is definitely preferable to ending up with a stale
dependency.