transactions: Add histogram to count voters
While we already track how many actions on a transaction are performed,
it's currently impossible to say how many voters are part of a single
transaction. This information may be quite important to asses current
state of a cluster.

Add a new histogram "gitaly_praefect_voters_per_transaction_total" to
keep track of this.