hook: Measure delay introduced by voting on transactions
While we already measure how long it takes until a transaction reaches
consensus on the server-side, we don't do so for the client-side. As
there may be additional latency involved due to having to connect to
Praefect first and doing an RPC, we should have both metrics.

Create a new metric "gitaly_reference_transaction_voting_delay_seconds"
to measure this on client-side.