praefect: Set up buckets for subtransaction histogram
We're currently using default buckets for the subtransactions histogram,
which are not really suited to display the number of subtransactions as
they include fractions, which cannot happen in the case of
subtransactions.

Fix this by defining our own buckets instead. We cannot use exponential
buckets here as we also want to include the zero-bucket.