Replication job acknowledge removes 'completed' and 'dead' events.
In order to prevent continues database growing we should remove
replication events that are moved to the final state: 'completed'
or 'dead'. The best way to achieve this is to remove events on the
stage of acknowledgment. Once the events is processed and should be
moved into 'completed' state or if there is no more attempts to
process it and the target state is 'dead' the event is removed from
the database without changing status. In order to support event
collapsing similar scheduled replication events would be removed
as well as there won't be any effect in applying them after successful
execution of initial one.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2941