Local elector that is used in case failover is disabled
Local elector doesn't handle retrieval of the primary node properly.
As it is not started by the Mgr and there is no health checks
executed for each node. In suck case each node considered
unhealthy as it has 0 successful health checks.

Disabled elector is used in case failover is disabled.
It returns a first passed in node as a primary and all the others
as secondaries. It also returns `ErrPrimaryNotHealthy` as other
electors in case primary is not healthy and can't serve the requests.

The health check request starts despite of failover for any type of
elector. Each node shows actual health status with `IsHealthy` method.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3011