git-hooks: Remove reference-transaction hook
There's a bug in the reference-transaction implementation in git-core
which will cause interleaving calls to the reference-transaction hook to
execute another hook by accident. While a fix has been proposed
upstream, the earliest it may be available is with Git v2.28.1. So let's
remove our reference-transaction hook for now to not expose this broken
behaviour.