Don't expand filesystem paths of wiki pages
Paths starting with a `~` tilde character were getting expanded by the
call to `File.expand_path` in `BlobEntry.normalize_dir`. This can cause
an exception when the tilde is followed by an invalid username, which
makes the whole wiki unusable.

This overrides `BlobEntry.normalize_dir` so it doesn't expand tildes
anymore, and in our case we also don't need to handle symlinks or
Windows paths.