hook: Fix transaction voting delay metric
The delay metric for voting on a transaction always gets recorded when
executing `voteOnTransaction()`. This is wrong, as this function will
always get executed in case there is no vote being performed due to
missing metadata. As a result, we'll perform measurements on this metric
all the time even if there's no transaction vote being performed, making
the metric essentially useless.

Fix the issue by moving the observation further down so it only gets
recorded in case we're actually performing a vote.