featureflags: Default-enable primary-wins reference transactions
As a first step towards enabling strong consistency by default, this
change default-enables reference transactions with the primary-wins
strategy. This is the smallest step we can make towards strong
consistency, as it will exercise the voting logic while the transaction
will always succeed for the primary.

So let's change both feature flags to default-on and update the
documentation to reflect that change.