transactions: Log transaction state when cancelling them
In order to make a transaction's outcome more visible, add a new log
message that is written when cancelling a transaction. Next to the
actual notification that it's got wrapped up, the log message will also
include the number of total nodes registered as part of the transaction,
how many committed and the number of subtransactions created. This
improves visibility and makes it a lot easier to reason about what the
actual result of a transaction is.