Prune objects older then 24h on garbage collection
By default the dangling objects remain in the object storage
for 2 weeks. This could be an issue if there are large files
or sensitive data being removed, but not fully deleted from
storage. The new configuration parameter `prune` for
`GarbageCollect` RPC results to execution of `gc` with additional
configuration option for `prune` command executed as part of
the garbage collection process. It removes all dangling objects
from the storage that are older then 24h.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/2919