Use repository generations to determine the best leader to elect
We should take into account repositories that are not yet exist
on the storage, so we should treat them as completely out of date.
With this we won't have situation when storage will be chosen as
primary because it has only one repository that is up to date, but
missing all other repositories that exist on other storages.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3008