coordinator: Setup metrics in coordinator
Right now, setup of metrics used in the coordinator is split across
multiple locations. This makes the process of adding new metrics more
involved than it needs to be and is a source of bugs in case any of
those locations is not updated.

Improve the situation by moving setup of metrics into the coordinator.
Metrics are exposed by implementing the Collector interface and
registering the coordinator itself as a metric.