gitaly-hooks: Pass through transaction information to post-receive hook
When using transactions and streaming commands to multiple nodes at
once, we need to avoid running Git hooks on all nodes taking part in the
transaction. To do so, the transactions contains a "primary" flag to
tell the given node that it is considerd the primary and should thus
execute the hook. While we already check for this in the post-receive
hook, we don't actually inject the transaction information. As a result,
each node won't find any transaction information, assume there is on
transaction and ultimately execute hooks even on nodes considered a
secondary.

Fix this by injecting tranasction information into the hook's
environment.