hooks: Honor primary-flag in Go post-receive hook
The Ruby hook logic already honors the primary-flag of transaction
information as it should, but the Go-implementation which is hidden
behind a feature flag doesn't. This would lead us to execute hooks on
secondaries if the Go implementation was enabled.

Fix this by conditionally executing the hook logic in Go code.