tests: Fix hook scripts using Git from PATH
In various tests we're setting up hooks in order to verify certain
operations behave as we intend them to. Some of those hooks we have
invoke Git again, but they'll use the PATH variable to do so. As a
result, they may use the wrong Git executable instead of the configured
one.

Fix this by generating hook scripts with the configured and resolved Git
path.