Makefile: Fix redirection of exit code
As we want to execute the JUnit testreport regardless of whether our Go
tests failed or not, we temporarily store the test's exit code in a
file. The redirection syntax to initialize that file is wrong though,
causing it to contain nothing at all. This accidentally does the right
thing as `exit $(cat file)` will exit with a zero return code if the
file is empty, but it does cause an annoying empty line to be emitted
when executing the Makefile target.

Remove this annoying newline by fixing redirection syntax.