Makefile: Inline check-mod-tidy
With PATH now containing a broken version of Git by design, we need to
adjust all callers of Git to use the real executable instead of the
broken one. The script `check-mod-tidy` is one of those which needs
adjusting.

Given it's such a tiny script, let's just inline it into our Makefile
and have it use the `${GIT}` variable.